+++ src/Web/Sort.ts
@@ -1,5 +1,5 @@
 /**
- * Specifies the order of a sort parameter.
+ * Specifies the direction of a sorted property.
  */
 export const SortDirection = Object.freeze({
 
@@ -15,14 +15,14 @@ export const SortDirection = Object.freeze({
 });
 
 /**
- * Specifies the order of a sort parameter.
+ * Specifies the direction of a sorted property.
  */
 export type SortDirection = typeof SortDirection[keyof typeof SortDirection];
 
 /**
- * Represents an attribute/order pair of a sort.
+ * Holds the name of a property and the direction to sort by.
  */
-export type SortEntry = [string, SortDirection];
+export type SortedProperty = [string, SortDirection];
 
 /**
  * Represents information relevant to the sorting of data items.
@@ -30,33 +30,33 @@ export type SortEntry = [string, SortDirection];
 export class Sort {
 
 	/**
-	 * The list of attribute/order pairs.
+	 * The list of sorted properties.
 	 */
-	#attributes: SortEntry[];
+	#properties: SortedProperty[];
 
 	/**
 	 * Creates new sort.
-	 * @param attributes The list of attributes to be sorted.
+	 * @param properties The list of properties to be sorted.
 	 */
-	constructor(attributes: SortEntry[] = []) {
-		this.#attributes = attributes;
+	constructor(properties: SortedProperty[] = []) {
+		this.#properties = properties;
 	}
 
 	/**
-	 * The number of attributes.
+	 * The number of properties.
 	 */
 	get length(): number {
-		return this.#attributes.length;
+		return this.#properties.length;
 	}
 
 	/**
-	 * Creates a new sort from the specified attribute and order.
-	 * @param attribute The sort attribute.
-	 * @param order THe sort order.
-	 * @returns The sort corresponding to the attribute and order.
+	 * Creates a new sort from the specified property and direction.
+	 * @param property The property name.
+	 * @param direction The sort direction.
+	 * @returns The sort corresponding to the property and direction.
 	 */
-	static of(attribute: string, order: SortDirection = SortDirection.Ascending): Sort {
-		return new this([[attribute, order]]);
+	static of(property: string, direction: SortDirection = SortDirection.Ascending): Sort {
+		return new this([[property, direction]]);
 	}
 
 	/**
@@ -66,82 +66,82 @@ export class Sort {
 	 */
 	static parse(value: string): Sort {
 		return new this((value ? value.split(",") : []).map(item => {
-			const order = item.startsWith("-") ? SortDirection.Descending : SortDirection.Ascending;
-			return [order == SortDirection.Ascending ? item : item.slice(1), order];
+			const direction = item.startsWith("-") ? SortDirection.Descending : SortDirection.Ascending;
+			return [direction == SortDirection.Ascending ? item : item.slice(1), direction];
 		}));
 	}
 
 	/**
 	 * Returns a new iterator that allows iterating the entries of this sort.
-	 * @returns An iterator over the attribute/order pairs.
+	 * @returns An iterator over the sorted properties.
 	 */
-	*[Symbol.iterator](): IterableIterator<SortEntry> {
-		for (const entry of this.#attributes) yield entry;
+	*[Symbol.iterator](): IterableIterator<SortedProperty> {
+		for (const entry of this.#properties) yield entry;
 	}
 
 	/**
-	 * Appends the specified attribute to this sort.
-	 * @param attribute The attribute name.
-	 * @param order The sort order.
+	 * Appends the specified property to this sort.
+	 * @param property The property name.
+	 * @param direction The sort direction.
 	 * @returns This instance.
 	 */
-	append(attribute: string, order: SortDirection): this {
-		this.delete(attribute);
-		this.#attributes.push([attribute, order]);
+	append(property: string, direction: SortDirection): this {
+		this.delete(property);
+		this.#properties.push([property, direction]);
 		return this;
 	}
 
 	/**
-	 * Gets the attribute/order pair at the specified index.
+	 * Gets the sorted property at the specified index.
 	 * @param index The position in this sort.
-	 * @returns The attribute/order pair at the specified index, or `null` if it doesn't exist.
+	 * @returns The sorted property at the specified index, or `null` if it doesn't exist.
 	 */
-	at(index: number): SortEntry|null {
-		return this.#attributes.at(index) ?? null;
+	at(index: number): SortedProperty|null {
+		return this.#properties.at(index) ?? null;
 	}
 
 	/**
-	 * Compares the specified objects, according to the current sort attributes.
+	 * Compares the specified objects, according to the current sort properties.
 	 * @param x The first object to compare.
 	 * @param y The second object to compare.
 	 * @returns A value indicating the relationship between the two objects.
 	 */
 	compare(x: object, y: object): number {
-		for (const [attribute, order] of this.#attributes) {
-			const xAttr = Reflect.get(x, attribute); // eslint-disable-line @typescript-eslint/no-unsafe-assignment
-			const yAttr = Reflect.get(y, attribute); // eslint-disable-line @typescript-eslint/no-unsafe-assignment
+		for (const [property, direction] of this.#properties) {
+			const xAttr = Reflect.get(x, property); // eslint-disable-line @typescript-eslint/no-unsafe-assignment
+			const yAttr = Reflect.get(y, property); // eslint-disable-line @typescript-eslint/no-unsafe-assignment
 			const value = xAttr > yAttr ? 1 : (xAttr < yAttr ? -1 : 0);
-			if (value) return order == SortDirection.Ascending ? value : -value;
+			if (value) return direction == SortDirection.Ascending ? value : -value;
 		}
 
 		return 0;
 	}
 
 	/**
-	 * Removes the specified attribute from this sort.
-	 * @param attribute The attribute name.
+	 * Removes the specified property from this sort.
+	 * @param property The property name.
 	 */
-	delete(attribute: string): void {
-		this.#attributes = this.#attributes.filter(([key]) => key != attribute);
+	delete(property: string): void {
+		this.#properties = this.#properties.filter(([key]) => key != property);
 	}
 
 	/**
-	 * Gets the order associated with the specified attribute.
-	 * @param attribute The attribute name.
-	 * @returns The order associated with the specified attribute, or `null` if the attribute doesn't exist.
+	 * Gets the direction associated with the specified property.
+	 * @param property The property name.
+	 * @returns The direction associated with the specified property, or `null` if the property doesn't exist.
 	 */
-	get(attribute: string): SortDirection|null {
-		for (const [key, order] of this.#attributes) if (key == attribute) return order;
+	get(property: string): SortDirection|null {
+		for (const [key, direction] of this.#properties) if (key == property) return direction;
 		return null;
 	}
 
 	/**
-	 * Gets the icon corresponding to the specified attribute.
-	 * @param attribute The attribute name.
-	 * @returns The icon corresponding to the specified attribute.
+	 * Gets the icon corresponding to the specified property.
+	 * @param property The property name.
+	 * @returns The icon corresponding to the specified property.
 	 */
-	getIcon(attribute: string): string {
-		switch (this.get(attribute)) {
+	getIcon(property: string): string {
+		switch (this.get(property)) {
 			case SortDirection.Ascending: return "arrow_upward";
 			case SortDirection.Descending: return "arrow_downward";
 			default: return "swap_vert";
@@ -149,33 +149,33 @@ export class Sort {
 	}
 
 	/**
-	 * Returns a value indicating whether the specified attribute exists in this sort.
-	 * @param attribute The attribute name.
-	 * @returns `true` if the specified attribute exists in this sort, otherwise `false`.
+	 * Returns a value indicating whether the specified property exists in this sort.
+	 * @param property The property name.
+	 * @returns `true` if the specified property exists in this sort, otherwise `false`.
 	 */
-	has(attribute: string): boolean {
-		return this.#attributes.some(([key]) => key == attribute);
+	has(property: string): boolean {
+		return this.#properties.some(([key]) => key == property);
 	}
 
 	/**
-	 * Gets the index of the specified attribute in the underlying list.
-	 * @param attribute The attribute name.
-	 * @returns The index of the specified attribute, or `-1` if the attribute is not found.
+	 * Gets the index of the specified property in the underlying list.
+	 * @param property The property name.
+	 * @returns The index of the specified property, or `-1` if the property is not found.
 	 */
-	indexOf(attribute: string): number {
-		for (const [index, [key]] of this.#attributes.entries()) if (key == attribute) return index;
+	indexOf(property: string): number {
+		for (const [index, [key]] of this.#properties.entries()) if (key == property) return index;
 		return -1;
 	}
 
 	/**
-	 * Prepends the specified attribute to this sort.
-	 * @param attribute The attribute name.
-	 * @param order The sort order.
+	 * Prepends the specified property to this sort.
+	 * @param property The property name.
+	 * @param direction The sort direction.
 	 * @returns This instance.
 	 */
-	prepend(attribute: string, order: SortDirection): this {
-		this.delete(attribute);
-		this.#attributes.unshift([attribute, order]);
+	prepend(property: string, direction: SortDirection): this {
+		this.delete(property);
+		this.#properties.unshift([property, direction]);
 		return this;
 	}
 
@@ -184,30 +184,30 @@ export class Sort {
 	 * @param conditions The conditions to satisfy.
 	 * @returns `true` if the current sort satisfies the specified conditions, otherwise `false`.
 	 */
-	satisfies(conditions: Partial<{attributes: string[], min: number, max: number}> = {}): boolean {
+	satisfies(conditions: Partial<{min: number, max: number, properties: string[]}> = {}): boolean {
 		const min = conditions.min ?? -1;
 		if (min >= 0) return this.length >= min;
 
 		const max = conditions.max ?? -1;
 		if (max >= 0) return this.length <= max;
 
-		const attributes = conditions.attributes ?? [];
-		return attributes.length ? this.#attributes.every(([key]) => attributes.includes(key)) : true;
+		const properties = conditions.properties ?? [];
+		return properties.length ? this.#properties.every(([key]) => properties.includes(key)) : true;
 	}
 
 	/**
-	 * Sets the order of the specified attribute.
-	 * @param attribute The attribute name.
-	 * @param order The sort order.
+	 * Sets the direction of the specified property.
+	 * @param property The property name.
+	 * @param direction The sort direction.
 	 * @returns This instance.
 	 */
-	set(attribute: string, order: SortDirection): this {
-		for (const [index, [key]] of this.#attributes.entries()) if (key == attribute) {
-			this.#attributes[index] = [key, order];
+	set(property: string, direction: SortDirection): this {
+		for (const [index, [key]] of this.#properties.entries()) if (key == property) {
+			this.#properties[index] = [key, direction];
 			return this;
 		}
 
-		return this.append(attribute, order);
+		return this.append(property, direction);
 	}
 
 	/**
@@ -224,7 +224,7 @@ export class Sort {
 	 * @returns The SQL clause corresponding to this object.
 	 */
 	toSql(escape: (identifier: string) => string = id => id): string {
-		return this.#attributes.map(([attribute, order]) => `${escape(attribute)} ${order}`).join(", ");
+		return this.#properties.map(([property, direction]) => `${escape(property)} ${direction}`).join(", ");
 	}
 
 	/**
@@ -232,6 +232,6 @@ export class Sort {
 	 * @returns The string representation of this object.
 	 */
 	toString(): string {
-		return this.#attributes.map(([attribute, order]) => `${order == SortDirection.Descending ? "-" : ""}${attribute}`).join();
+		return this.#properties.map(([property, direction]) => `${direction == SortDirection.Descending ? "-" : ""}${property}`).join();
 	}
 }